requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.0.1
django-cors-headers==4.3.1
requests==2.31.0
geopy==2.4.0
twilio==8.10.0
opencv-python==4.8.1.78
face-recognition==1.3.0